requirements.txt
requests==2.28.1
django==4.0.5
django-colorfield==0.7.1
# psycopg2==2.9.3    # если выдает ошибку
# pip3 install psycopg2==2.9.3
python-dotenv==0.20.0
